+++ extensions/data-plane/data-plane-api/src/main/java/org/eclipse/dataspaceconnector/dataplane/api/DataPlaneApiExtension.java
@@ -14,8 +14,9 @@
 package org.eclipse.dataspaceconnector.dataplane.api;
 
 import okhttp3.OkHttpClient;
-import org.eclipse.dataspaceconnector.dataplane.api.transfer.DataPlanePublicApiRequestFilter;
-import org.eclipse.dataspaceconnector.dataplane.api.transfer.DataPlaneTransferController;
+import org.eclipse.dataspaceconnector.dataplane.api.controller.ContainerRequestContextApiImpl;
+import org.eclipse.dataspaceconnector.dataplane.api.controller.DataPlanePublicApiController;
+import org.eclipse.dataspaceconnector.dataplane.api.controller.DataPlaneTransferController;
 import org.eclipse.dataspaceconnector.dataplane.api.validation.RemoteTokenValidationService;
 import org.eclipse.dataspaceconnector.dataplane.spi.manager.DataPlaneManager;
 import org.eclipse.dataspaceconnector.spi.EdcSetting;
@@ -53,7 +54,9 @@ public class DataPlaneApiExtension implements ServiceExtension {
         var controlPlaneAddress = context.getSetting(CONTROL_PLANE_VALIDATION_ENDPOINT, "/api/validation");
         var tokenValidationClient = new RemoteTokenValidationService(httpClient, controlPlaneAddress, context.getTypeManager().getMapper());
         webService.registerResource(CONTROL, new DataPlaneTransferController(dataPlaneManager));
-        webService.registerResource(PUBLIC, new DataPlanePublicApiRequestFilter(tokenValidationClient, dataPlaneManager, context.getMonitor(), context.getTypeManager()));
+
+        var publicApiController = new DataPlanePublicApiController(dataPlaneManager, tokenValidationClient, context.getMonitor(), new ContainerRequestContextApiImpl(), context.getTypeManager());
+        webService.registerResource(PUBLIC, publicApiController);
     }
 }