+++ CHANGELOG.md
@@ -27,6 +27,7 @@ the detailed section referring to by linking pull requests or issues.
 * Implement S3BucketReader (#675)
 * Add configuration setting for state machine batch size (#872)
 * Add Jetty context alias for IDS API (#815)
+* Pass path information into http data source through DPF public API (#929)
 
 #### Changed
 
@@ -49,8 +50,9 @@ the detailed section referring to by linking pull requests or issues.
 * Remove module `:samples:other:streaming` (#889)
 
 #### Fixed
+
 * Flaky S3 StatusChecker Test (#794)
-* Added missing Data Management Asset controller openapi (#853) 
+* Added missing Data Management Asset controller openapi (#853)
 
 ---